// src/enums/websocket.ts
export enum WebsocketEvents {
    TEAM_JOINED = 'TEAM_JOINED',
    TEAM_ANSWERED = 'TEAM_ANSWERED',
    SCOREBOARD_CONNECTED = 'SCOREBOARD_CONNECTED',
    TEAM_ACCEPTED = 'TEAM_ACCEPTED',
    TEAM_DECLINED = 'TEAM_DECLINED',
    NEW_QUESTION = 'NEW_QUESTION',
    ROUND_FINISHED = 'ROUND_FINISHED',
    QUIZ_ENDED = 'QUIZ_ENDED',
    QUESTION_CLOSED = 'QUESTION_CLOSED',
    QUESTION_APPROVED = 'QUESTION_APPROVED'
}